refactor: interface의 증가에 따라 각 컴포넌트 패키지 이하로 interface 이동
