+++ app/src/main/java/com/codesoom/assignment/controllers/ToyCrudController.java
@@ -3,8 +3,8 @@ package com.codesoom.assignment.controllers;
 import com.codesoom.assignment.controllers.dtos.ToyRequestDto;
 import com.codesoom.assignment.controllers.dtos.ToyResponseDto;
 import com.codesoom.assignment.domain.Toy;
-import com.codesoom.assignment.interfaces.ProductCrudController;
-import com.codesoom.assignment.interfaces.ProductCrudService;
+import com.codesoom.assignment.controllers.interfaces.ProductCrudController;
+import com.codesoom.assignment.application.interfaces.ProductCrudService;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
