+++ app/src/main/java/com/codesoom/assignment/controllers/interfaces/ProductCrudController.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.interfaces;
+package com.codesoom.assignment.controllers.interfaces;
 
 import com.codesoom.assignment.controllers.dtos.ToyRequestDto;
 import com.codesoom.assignment.controllers.dtos.ToyResponseDto;
