+++ app/src/main/java/com/codesoom/assignment/domain/interfaces/MediaDemo.java
@@ -1,0 +1,5 @@
+package com.codesoom.assignment.domain.interfaces;
+
+public interface MediaDemo {
+    String url();
+}
