+++ app/src/main/java/com/codesoom/assignment/domain/interfaces/Money.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.interfaces;
+package com.codesoom.assignment.domain.interfaces;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
