+++ app/src/main/java/com/codesoom/assignment/domain/interfaces/Producer.java
@@ -1,0 +1,5 @@
+package com.codesoom.assignment.domain.interfaces;
+
+public interface Producer {
+    String name();
+}
