+++ app/src/main/java/com/codesoom/assignment/domain/interfaces/Shop.java
@@ -1,0 +1,5 @@
+package com.codesoom.assignment.domain.interfaces;
+
+public interface Shop {
+    String name();
+}
