--- app/src/main/java/com/codesoom/assignment/interfaces/MediaDemo.java
@@ -1,5 +1,0 @@
-package com.codesoom.assignment.interfaces;
-
-public interface MediaDemo {
-    String url();
-}
