--- app/src/main/java/com/codesoom/assignment/interfaces/Producer.java
@@ -1,5 +1,0 @@
-package com.codesoom.assignment.interfaces;
-
-public interface Producer {
-    String name();
-}
